moves ai context client to mcp
